Remove error when not in WP context
Some tools (such as [php-composter](https://github.com/php-composter/php-composter)) load the `vendor/autoload.php` from the command line. In this context WP is not available and the previous version would crash on an `undefined function add_action` or similar. The WP context is not needed to do static checking such as phpMD or phpCS.