Skip class_exists in non-debug settings
Calling class_exists would trigger a lookup from composer to find the corresponding hook file, even if it is not defined in the classmap file generated by composer.

In non debugging modes we look up the file in the classmap. If it does not exists we immedietly return without calling class_exists. This results in many less file_exists calls by Composer.